added token count to hero; added weapon type check when equipping weapons.

+++ Assets/Scripts/Avatars/HeroData.cs
@@ -7,7 +7,7 @@ public class HeroData : AvatarData
 {
    public int level = 1;        //can be used to track where hero should be in the xp table
    public int MaxLevel {get;} = 50;
-   public int attackTokenMod;    //adjusts how many attacks a hero gets. Can be negative, but minimum attack count is 1.
+   public short attackTokenMod;    //adjusts how many attacks a hero gets. Can be negative, but minimum attack count is 1.
    [Header("Starting gear")]
     public Weapon weapon;
     public Armor armor;
